+++ examples/standalone/expo-linearlite/src/components/Filters.tsx
@@ -8,7 +8,7 @@ export const Filters: React.FC = () => {
   const { store } = useStore()
   const [{ filter }] = useRow(tables.app)
 
-  const setFilter = (filter: Filter) => store.mutate(mutations.setFilter({ filter }))
+  const setFilter = (filter: Filter) => store.mutate(mutations.filterUpdated({ filter }))
 
   return (
     <View style={styles.container}>