+++ examples/standalone/todomvc-experimental/src/components/Footer.tsx
@@ -15,7 +15,7 @@ export const Footer: React.FC = () => {
   const [{ filter }] = useRow(tables.app, sessionId)
   const incompleteCount = useQuery(incompleteCount$)
 
-  const setFilter = (filter: Filter) => store.mutate(mutations.setFilter({ filter, sessionId }))
+  const setFilter = (filter: Filter) => store.mutate(mutations.filterUpdated({ filter, sessionId }))
 
   return (
     <footer className="footer">
@@ -39,7 +39,7 @@ export const Footer: React.FC = () => {
       </ul>
       <button
         className="clear-completed"
-        onClick={() => store.mutate(mutations.clearCompleted({ deleted: Date.now() }))}
+        onClick={() => store.mutate(mutations.todoClearedCompleted({ deleted: Date.now() }))}
       >
         Clear completed
       </button>