+++ packages/@livestore/common/src/sync/next/test/compact-events.test.ts
@@ -81,17 +81,17 @@ const compact = (events: any[]) => {
 }
 
 describe('compactEvents todo app', () => {
-  it('completeTodo', () => {
+  it('todoCompleted', () => {
     const expected = compact([
       mutations.createTodo({ id: 'A', text: 'buy milk' }), // 0
-      mutations.completeTodo({ id: 'A' }), // 1
-      mutations.completeTodo({ id: 'A' }), // 2
+      mutations.todoCompleted({ id: 'A' }), // 1
+      mutations.todoCompleted({ id: 'A' }), // 2
     ])
 
     expect(expected).toMatchInlineSnapshot(`
       [
         { id: 0, parentId: -1, mutation: "createTodo", args: { id: "A", text: "buy milk" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-A +todo-is-writeable-A=true +todo-completed-A=false" }
-        { id: 2, parentId: 0, mutation: "completeTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
+        { id: 2, parentId: 0, mutation: "todoCompleted", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
       ]
     `)
   })
@@ -114,20 +114,20 @@ describe('compactEvents todo app', () => {
     `)
   })
 
-  it('completeTodo / toggleTodo', () => {
+  it('todoCompleted / toggleTodo', () => {
     const expected = compact([
       mutations.createTodo({ id: 'A', text: 'buy milk' }), // 0
       mutations.toggleTodo({ id: 'A' }), // 1
       mutations.toggleTodo({ id: 'A' }), // 2
-      mutations.completeTodo({ id: 'A' }), // 3
-      mutations.completeTodo({ id: 'A' }), // 4
+      mutations.todoCompleted({ id: 'A' }), // 3
+      mutations.todoCompleted({ id: 'A' }), // 4
       mutations.toggleTodo({ id: 'A' }), // 5
     ])
 
     expect(expected).toMatchInlineSnapshot(`
       [
         { id: 0, parentId: -1, mutation: "createTodo", args: { id: "A", text: "buy milk" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-A +todo-is-writeable-A=true +todo-completed-A=false" }
-        { id: 4, parentId: 0, mutation: "completeTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
+        { id: 4, parentId: 0, mutation: "todoCompleted", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
         { id: 5, parentId: 4, mutation: "toggleTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true ?todo-completed-A +todo-completed-A=false" }
       ]
     `)
@@ -155,7 +155,7 @@ describe('compactEvents todo app', () => {
     const expected = compact([
       mutations.createTodo({ id: 'A', text: 'buy milk' }), // 0
       mutations.setReadonlyTodo({ id: 'A', readonly: false }), // 1
-      mutations.completeTodo({ id: 'A' }), // 2
+      mutations.todoCompleted({ id: 'A' }), // 2
       mutations.setTextTodo({ id: 'A', text: 'buy soy milk' }), // 3
       mutations.setReadonlyTodo({ id: 'A', readonly: true }), // 4
     ])
@@ -164,7 +164,7 @@ describe('compactEvents todo app', () => {
       [
         { id: 0, parentId: -1, mutation: "createTodo", args: { id: "A", text: "buy milk" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-A +todo-is-writeable-A=true +todo-completed-A=false" }
         { id: 1, parentId: 0, mutation: "setReadonlyTodo", args: { id: "A", readonly: false }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A +todo-is-writeable-A=true" }
-        { id: 2, parentId: 1, mutation: "completeTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
+        { id: 2, parentId: 1, mutation: "todoCompleted", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
         { id: 3, parentId: 2, mutation: "setTextTodo", args: { id: "A", text: "buy soy milk" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-text-updated-A" }
         { id: 4, parentId: 3, mutation: "setReadonlyTodo", args: { id: "A", readonly: true }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A +todo-is-writeable-A=false" }
       ]
@@ -188,14 +188,14 @@ describe('compactEvents todo app', () => {
     `)
   })
 
-  it('completeTodos', () => {
+  it('todoCompleteds', () => {
     const expected = compact([
       mutations.createTodo({ id: 'A', text: 'buy milk' }), // 0
       mutations.createTodo({ id: 'B', text: 'buy bread' }), // 1
       mutations.createTodo({ id: 'C', text: 'buy cheese' }), // 2
-      mutations.completeTodos({ ids: ['A', 'B', 'C'] }), // 3
+      mutations.todoCompleteds({ ids: ['A', 'B', 'C'] }), // 3
       mutations.toggleTodo({ id: 'A' }), // 4
-      mutations.completeTodo({ id: 'A' }), // 5
+      mutations.todoCompleted({ id: 'A' }), // 5
     ])
 
     expect(expected).toMatchInlineSnapshot(`
@@ -203,21 +203,21 @@ describe('compactEvents todo app', () => {
         { id: 0, parentId: -1, mutation: "createTodo", args: { id: "A", text: "buy milk" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-A +todo-is-writeable-A=true +todo-completed-A=false" }
         { id: 1, parentId: 0, mutation: "createTodo", args: { id: "B", text: "buy bread" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-B +todo-is-writeable-B=true +todo-completed-B=false" }
         { id: 2, parentId: 1, mutation: "createTodo", args: { id: "C", text: "buy cheese" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-C +todo-is-writeable-C=true +todo-completed-C=false" }
-        { id: 3, parentId: 2, mutation: "completeTodos", args: { ids: ["A", "B", "C"] }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true ↖todo-exists-B ↖todo-is-writeable-B=true ↖todo-exists-C ↖todo-is-writeable-C=true +todo-completed-A=true +todo-completed-B=true +todo-completed-C=true" }
-        { id: 5, parentId: 3, mutation: "completeTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
+        { id: 3, parentId: 2, mutation: "todoCompleteds", args: { ids: ["A", "B", "C"] }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true ↖todo-exists-B ↖todo-is-writeable-B=true ↖todo-exists-C ↖todo-is-writeable-C=true +todo-completed-A=true +todo-completed-B=true +todo-completed-C=true" }
+        { id: 5, parentId: 3, mutation: "todoCompleted", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
       ]
     `)
   })
 
-  it('completeTodos 2', () => {
+  it('todoCompleteds 2', () => {
     const expected = compact([
       mutations.createTodo({ id: 'A', text: 'buy milk' }), // 0
       mutations.createTodo({ id: 'B', text: 'buy bread' }), // 1
       mutations.createTodo({ id: 'C', text: 'buy cheese' }), // 2
       mutations.toggleTodo({ id: 'A' }), // 3
-      mutations.completeTodos({ ids: ['A', 'B', 'C'] }), // 4
+      mutations.todoCompleteds({ ids: ['A', 'B', 'C'] }), // 4
       mutations.toggleTodo({ id: 'A' }), // 5
-      mutations.completeTodo({ id: 'A' }), // 6
+      mutations.todoCompleted({ id: 'A' }), // 6
     ])
 
     expect(expected).toMatchInlineSnapshot(`
@@ -225,8 +225,8 @@ describe('compactEvents todo app', () => {
         { id: 0, parentId: -1, mutation: "createTodo", args: { id: "A", text: "buy milk" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-A +todo-is-writeable-A=true +todo-completed-A=false" }
         { id: 1, parentId: 0, mutation: "createTodo", args: { id: "B", text: "buy bread" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-B +todo-is-writeable-B=true +todo-completed-B=false" }
         { id: 2, parentId: 1, mutation: "createTodo", args: { id: "C", text: "buy cheese" }, clientId: "client-id", sessionId: "session-id", facts: "+todo-exists-C +todo-is-writeable-C=true +todo-completed-C=false" }
-        { id: 4, parentId: 2, mutation: "completeTodos", args: { ids: ["A", "B", "C"] }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true ↖todo-exists-B ↖todo-is-writeable-B=true ↖todo-exists-C ↖todo-is-writeable-C=true +todo-completed-A=true +todo-completed-B=true +todo-completed-C=true" }
-        { id: 6, parentId: 4, mutation: "completeTodo", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
+        { id: 4, parentId: 2, mutation: "todoCompleteds", args: { ids: ["A", "B", "C"] }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true ↖todo-exists-B ↖todo-is-writeable-B=true ↖todo-exists-C ↖todo-is-writeable-C=true +todo-completed-A=true +todo-completed-B=true +todo-completed-C=true" }
+        { id: 6, parentId: 4, mutation: "todoCompleted", args: { id: "A" }, clientId: "client-id", sessionId: "session-id", facts: "↖todo-exists-A ↖todo-is-writeable-A=true +todo-completed-A=true" }
       ]
     `)
   })